Ensure locale directive is updated client-side
